Alex and Jamie feedback/requirements
