+++ include/constants/items.h
@@ -1140,4 +1140,8 @@ enum EffectItem
     EFFECT_ITEM_USE_POKE_FLUTE,
 };
 
+#define ITEM_RANDOM_OPTION_0 6000
+#define ITEM_RANDOM_OPTION_1 6001
+#define ITEM_RANDOM_OPTION_2 6002
+
 #endif  // GUARD_CONSTANTS_ITEMS_H
