+++ include/constants/pokeball.h
@@ -34,4 +34,6 @@ enum PokeBall
     POKEBALL_COUNT
 };
 
+#define BALL_RANDOM (POKEBALL_COUNT + 1)
+
 #endif // GUARD_CONSTANTS_POKEBALL_H
